clean up database logic
