+++ internal/persistence/database/database.go
@@ -1,0 +1,65 @@
+package database
+
+import (
+	"context"
+	"database/sql"
+	"errors"
+)
+
+type Database struct {
+	db *sql.DB
+}
+
+func New(db *sql.DB) Database {
+	return Database{db: db}
+}
+
+type stmtHandler interface {
+	ExecContext(context.Context, string, ...any) (sql.Result, error)
+	QueryContext(context.Context, string, ...any) (*sql.Rows, error)
+	QueryRowContext(context.Context, string, ...any) *sql.Row
+}
+
+func getDB(ctx context.Context, db *sql.DB) stmtHandler {
+	tx, ok := getTx(ctx)
+	if ok {
+		return tx
+	}
+	return db
+}
+
+func (db Database) ExecContext(ctx context.Context, stmt string, args ...any) (sql.Result, error) {
+	return getDB(ctx, db.db).ExecContext(ctx, stmt, args...)
+}
+
+func (db Database) ExecVoid(ctx context.Context, stmt string, args ...any) error {
+	_, err := getDB(ctx, db.db).ExecContext(ctx, stmt, args...)
+	return err
+}
+
+func (db Database) ExecOne(ctx context.Context, stmt string, args ...any) error {
+	res, err := getDB(ctx, db.db).ExecContext(ctx, stmt, args...)
+	if err != nil {
+		return err
+	}
+
+	if n, err := res.RowsAffected(); err != nil {
+		return err
+	} else if n != 1 {
+		return errors.New("expected exactly one row to be affected")
+	}
+
+	return nil
+}
+
+func (db Database) QueryContext(ctx context.Context, stmt string, args ...any) (*sql.Rows, error) {
+	return getDB(ctx, db.db).QueryContext(ctx, stmt, args...)
+}
+
+func (db Database) QueryRowContext(ctx context.Context, stmt string, args ...any) *sql.Row {
+	return getDB(ctx, db.db).QueryRowContext(ctx, stmt, args...)
+}
+
+func (db Database) WithTx(ctx context.Context, fn func(context.Context) error) error {
+	return withTx(ctx, db.db, fn)
+}
