+++ internal/persistence/divisions/divisions.go
@@ -3,7 +3,7 @@ package divisions
 import (
 	"context"
 
-	"github.com/cszczepaniak/cribbly/internal/persistence"
+	"github.com/cszczepaniak/cribbly/internal/persistence/database"
 	"github.com/cszczepaniak/go-sqlbuilder/sqlbuilder"
 	"github.com/cszczepaniak/go-sqlbuilder/sqlbuilder/column"
 	"github.com/cszczepaniak/go-sqlbuilder/sqlbuilder/filter"
@@ -19,11 +19,11 @@ type Division struct {
 }
 
 type Repository struct {
-	db persistence.Database
+	db database.Database
 	b  *sqlbuilder.Builder
 }
 
-func NewRepository(db persistence.Database) Repository {
+func NewRepository(db database.Database) Repository {
 	return Repository{
 		db: db,
 		b:  sqlbuilder.New(formatter.Sqlite{}),
