+++ internal/persistence/sqlite/sqlite.go
@@ -6,18 +6,24 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/cszczepaniak/cribbly/internal/persistence/database"
 	_ "github.com/ncruces/go-sqlite3/driver"
 	_ "github.com/ncruces/go-sqlite3/embed"
 )
 
-func New(dsn string) (*sql.DB, error) {
+func New(dsn string) (database.Database, error) {
 	filePath, ok := strings.CutPrefix(dsn, "file:")
 	if ok && filePath != ":memory:" {
 		err := os.MkdirAll(filepath.Dir(filePath), os.ModePerm)
 		if err != nil {
-			return nil, err
+			return database.Database{}, err
 		}
 	}
 
-	return sql.Open("sqlite3", dsn)
+	db, err := sql.Open("sqlite3", dsn)
+	if err != nil {
+		return database.Database{}, err
+	}
+
+	return database.New(db), nil
 }
