+++ internal/server/config.go
@@ -2,7 +2,7 @@ package server
 
 import (
 	"github.com/cszczepaniak/cribbly/internal/notifier"
-	"github.com/cszczepaniak/cribbly/internal/persistence"
+	"github.com/cszczepaniak/cribbly/internal/persistence/database"
 	"github.com/cszczepaniak/cribbly/internal/persistence/divisions"
 	"github.com/cszczepaniak/cribbly/internal/persistence/games"
 	"github.com/cszczepaniak/cribbly/internal/persistence/players"
@@ -13,7 +13,7 @@ import (
 )
 
 type Config struct {
-	Transactor          persistence.Transactor
+	Transactor          database.Transactor
 	PlayerRepo          players.Repository
 	TeamRepo            teams.Repository
 	DivisionRepo        divisions.Repository
